% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_rest.R
\name{publish_rest}
\alias{publish_rest}
\title{publish_rest}
\usage{
publish_rest(file, post_title, post_id = 0, status = "publish")
}
\arguments{
\item{file}{The path and file name of an R Markdown document.}

\item{post_title}{The title that should be used for the post.}

\item{post_id}{A post ID to update with the HTML content.}

\item{status}{The WordPress post status to assign. Default "publish".}
}
\value{
Nothing?
}
\description{
Renders an R Markdown document into an HTML fragment and
publishes the HTML as a post in WordPress using the WordPress REST API.
}
\examples{
publish_rest(file = 'current_file.Rmd', post_title = 'My post title')
}
