% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rest_credentials.R
\name{set_rest_credentials}
\alias{set_rest_credentials}
\title{set_rest_credentials}
\usage{
set_rest_credentials(
  site_url,
  site_user,
  site_password,
  path = "/wp-json/wp/v2/posts"
)
}
\arguments{
\item{site_url}{The URL of the site.}

\item{site_user}{The username to assign the content to.}

\item{site_password}{The application password created for this user.}

\item{path}{An optional path if different than the WordPress default.}
}
\description{
Stores a set of credentials to use when publishing content
             using the WordPress REST API.

An application password must be created in WordPress for the user who
is publishing the content.
}
\examples{
set_rest_credentials( site_url = 'https://example.com', site_user = 'WPUser', site_password = 'ABCD EFGH IJKL MNOP' )
}
